Grab BGRA framebuffer data as BGR
Some drivers provide BGRA framebuffers even if alpha is not requested.
For now we just convert and take the performance hit. Eventually we
should send either BGR or BGRA data, but first the perf hit from
conversion will need to be compared to the extra bandwidth of sending an
unnecessary alpha channel.