Track the last Blender binary used in a new .pman.user config file
This allows using a version of Blender that is not on the system PATH.
To revert to the old behavior, set use_last_path in .pman.user to False.
The .pman.user file is automatically created if it does not exist. This
is a per-user file and should not be put under version control.